move resource to retrieve all rides created by a user to appropriate file
